[wip] udev support with async tokio-udev
Support for listening asynchronously to udev events with the `tokio-udev` crate.

This does not work as-is. Unfortunately, sometimes udev events get 'stuck' and
only processed after another batch of events comes in. In practice, this
translates to e.g. a batch of 'Add' events (after connecting a device) only
passed to the application loop after a batch of 'Remove' events (after
disconnecting the same device for instance).