[wip] left and right mouse button support
Add basic support for the left and right mouse buttons. Mouse buttons can also
be combined with the keyboard modifier state, which is now shared with all
devices.

As a mouse is (usually) not recognized as a keyboard by swhkd, the correct
device name must be passed to swhkd with `--device`.